# src/main/resources/static/openapi/YONDO_api.yaml
openapi: 3.0.3
info:
  title: KCC API
  description: |
    RESTful API 설계 가이드에 맞춰 수정된 사용자 인증/정보 관련 API 명세
    (로그인, 로그아웃, 비밀번호 초기화, 회원가입, 회원정보 수정, 토큰 재발급 등)
    + 확장: 트레이너 배정 ON/OFF, 일일 이용권 안내, PT 스케줄 관리, 직원 공지사항
  version: "1.0.0"
servers:
  - url: http://202.31.147.127:9345
    description: 서버 (개발용)
paths:
...
  /api/users/{identifier}:
    patch:
      summary: "회원 정보 수정"
      description: |
        * 기본적으로 `password`, `refund_account`, `receiptInfo`, `refund_bank`만 수정 가능.
        * 관리자인 경우 더 많은 필드 수정 가능(예: identifier, name, phone 등).
        * 권한 없는 필드 수정 요청 시 403 오류.
      operationId: modifyInfo
      security:
        - bearerAuth: []
      parameters:
        - name: identifier
          in: path
          required: true
          description: "수정 대상 사용자 식별자 (ex: testUser)"
          schema:
            type: string
      requestBody:
        required: true
        description: "수정할 정보(JSON)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyInfoRequest"
            examples:
              normal:
                summary: "정상 수정 예시 (일반 사용자)"
                value:
                  password: "newPassword"
                  refund_account: "110-987-654321"
                  receiptInfo: false
                  refund_bank: "우리은행"
              adminEdit:
                summary: "관리자 권한으로 모든 필드를 수정하는 예시"
                value:
                  identifier: "adminCanChangeThis"
                  name: "관리자 수정된 이름"
                  phone: "010-9999-8888"
                  branch: "부산점"
      responses:
        "200":
          description: "수정 완료"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "수정이 완료되었습니다."
        "403":
          description: "권한이 없는 사용자가 제한 필드를 수정하려고 할 경우"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                noPermission:
                  summary: "필드 수정 권한 없음"
                  value:
                    status: 403
                    success: false
                    message: "권한이 없습니다"
        "500":
          $ref: "#/components/responses/InternalServerError"


  ########################################
  # 4) 비밀번호 초기화: PATCH /api/users/{identifier}/password
  ########################################
...
  /api/admin/modify-info:
    get:
      summary: "관리자가 회원들의 정보를 불러오는 API"
      description: |
        * 예: 전체 사용자 목록 조회, 또는 특정 조건에 맞는 사용자들 조회.
        * 관리자 권한 필요. (SecurityContext에서 ADMIN 여부 확인)
      operationId: adminGetUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "관리자용 사용자 목록 조회 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        identifier:
                          type: string
                          example: "testUser"
                        name:
                          type: string
                          example: "홍길동"
                        phone:
                          type: string
                          example: "010-1234-5678"
                        role:
                          type: string
                          example: "MEMBER"
                example:
                  status: 200
                  success: true
                  data:
                    - identifier: "testUser"
                      name: "홍길동"
                      phone: "010-1234-5678"
                      role: "MEMBER"
                    - identifier: "adminUser"
                      name: "관리자"
                      phone: "010-0000-1111"
                      role: "ADMIN"
        "403":
          description: "권한이 없는 사용자(ADMIN이 아님) 혹은 ban된 사용자"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            examples:
              noAdminPermission:
                summary: "admin 권한 없음"
                value:
                  status: 403
                  success: false
                  message: "권한이 없습니다(관리자 전용)"
        "500":
          $ref: "#/components/responses/InternalServerError"

  ########################################
  # 7) PATCH /api/admin/modify-info/{identifier}
  ########################################
  /api/admin/modify-info/{identifier}:
    patch:
      summary: "관리자가 특정 사용자의 정보를 수정"
      description: |
        * 관리자 권한 필요.
        * Path Param: 수정 대상 식별자
        * RequestBody: 수정할 필드들
      operationId: adminModifyUser
      security:
        - bearerAuth: []
      parameters:
        - name: identifier
          in: path
          required: true
          description: "수정 대상 사용자 식별자"
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyInfoRequest"
            examples:
              adminSample:
                summary: "관리자가 사용자 정보 수정 예시"
                value:
                  name: "수정된 이름"
                  phone: "010-9999-8888"
                  role: "TRAINER"
      responses:
        "200":
          description: "수정 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "수정 성공"
        "403":
          description: "권한이 없는 사용자(ADMIN이 아님) 혹은 ban된 사용자"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          examples:
            noAdminPermission:
              summary: "admin 권한 없음"
              value:
                status: 403
                success: false
                message: "권한이 없습니다(관리자 전용)"
        "404":
          description: "대상 사용자를 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          examples:
            userNotFound:
              summary: "존재하지 않는 식별자"
              value:
                status: 404
                success: false
                message: "해당 사용자를 찾을 수 없습니다"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: "관리자가 특정 사용자를 삭제"
      description: |
        * 관리자 권한 필요.
        * Path Param: 삭제 대상 식별자
      operationId: adminDeleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: identifier
          in: path
          required: true
          description: "삭제 대상 사용자 식별자"
          schema:
            type: string
      responses:
        "200":
          description: "삭제 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "사용자 삭제 완료"
        "403":
          description: "권한이 없는 사용자(ADMIN이 아님) 혹은 ban된 사용자"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            examples:
              noAdminPermission:
                summary: "admin 권한 없음"
                value:
                  status: 403
                  success: false
                  message: "권한이 없습니다(관리자 전용)"
        "404":
          description: "삭제 대상 사용자를 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            examples:
              userNotFound:
                summary: "존재하지 않는 식별자"
                value:
                  status: 404
                  success: false
                  message: "해당 사용자를 찾을 수 없습니다"
        "500":
          $ref: "#/components/responses/InternalServerError"


  ####################################################
  # 8) 트레이너 배정 On/Off 설정: PATCH /api/admin/trainer-assign
  ####################################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ModifyInfoRequest:
      type: object
      properties:
        password:
          type: string
          format: password
          example: "newPassword"
        refundAccount:
          type: string
          example: "110-987-654321"
        receiptInfo:
          type: string
          example: "발급"
        refundBank:
          type: string
          example: "우리은행"
        phone:
          type: string
          example: "010-9999-8888"
        name:
          type: string
          example: "수정된 이름"
        branch:
          type: string
          example: "부산점"
        role:
          type: string
          example: "TRAINER"
        # 관리자 권한 시 identifier 등 추가 필드 수정 가능
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 401
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "에러 메시지 내용"
        code:
          type: string
          example: "T001"
      description: "에러 응답 공통 구조"

  ########################################
  # 공통 응답 (서버 오류 등)
  ########################################
  responses:
    InternalServerError:
      description: "서버 내부 오류"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalError:
              summary: "서버 에러 예시"
              value:
                status: 500
                success: false
                message: "Internal server error"
                code: "E999"